feat(web): server-render discovery pages
